+++ src/main/java/com/example/YouTube/repository/PlaylistRepository.java
@@ -3,7 +3,7 @@ package com.example.YouTube.repository;
 import com.example.YouTube.entity.PlaylistEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlaylistRepository extends CrudRepository<PlaylistEntity, Long> {
+public interface PlaylistRepository extends CrudRepository<PlaylistEntity, Integer> {
 
 
 }
